Update readme for new Closure compiler home
Interesting that the compiler is now on Github :-) A minor update to point the link to the wiki FAQ on its new home.

There is a new version of the compiler here 20140508. Would it not be simpler to have `compiler_url` also download the latest rather than a specific version?